for now, only let admins delete servers

+++ src/bot/discord/discord.go
@@ -94,11 +94,23 @@ func (h *handler) messageCreate(session *discordgo.Session, message *discordgo.M
 }
 
 func (h *handler) handleMessage(ctx context.Context, session *discordgo.Session, message *discordgo.MessageCreate, cmd string, args string) {
+
+	p, err := session.State.MessagePermissions(message.Message)
+	if err != nil {
+		sendMessage(session, message, fmt.Sprintf("Error getting permissions: %v", err.Error()))
+
+		return
+	}
+
 	switch {
 	case cmd == CreateServer:
 		h.handleCreateServer(ctx, session, message, args)
 
 	case cmd == DeleteServer:
+		if p&discordgo.PermissionAdministrator == 0 {
+			sendMessage(session, message, "You don't have permission to delete servers")
+			return
+		}
 		h.handleDeleteServer(ctx, session, message, args)
 
 	case cmd == ListServer:
